+++ src/main/java/racing/domain/Car.java
@@ -1,4 +1,4 @@
-package racing;
+package racing.domain;
 
 import static utility.StringHelper.throwExceptionNullOrLengthGreaterThan;
 import static racing.RacingConstant.*;
