+++ src/main/java/racing/domain/Racing.java
@@ -1,5 +1,6 @@
-package racing;
+package racing.domain;
 
+import racing.view.Winner;
 import utility.NumberHelper;
 
 import java.util.*;
