+++ src/test/java/racing/RacingTest.java
@@ -1,5 +1,8 @@
 package racing;
 
+import racing.domain.Car;
+import racing.domain.Racing;
+import racing.view.Winner;
 import utility.NumberHelper;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
